improve the scraper: use blacklists and a list of user agents
* a movie or a user can now be blacklisted
* the blacklists are automatically filled with the scraper when a problem
occurs with a movie or a user
* three new options for the scraper: "movie-blacklist", "user-blacklist"
 and "user-agents"
* the blacklists are created by the scraper if the files don't exist, each
line of the files contain one CK movie id or one CK user id
* the user-agent file contains one user agent by line. A user agent is
randomly choosen for each request
* a new limit is defined for the number of ratings of one user. That
prevents to scrape some users that are not human user. User with
an unacceptable number of rating is automatically blacklisted by the
scraper
* some refactoring is done in the CkScraper: each request to the website
is done in the same method. Thus, the pause between two requests is
centralized.